requirements-dev.txt
# Requirements para desenvolvimento local
-r requirements.txt

# Ferramentas de desenvolvimento
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Jupyter para análise de dados (opcional)
jupyter==1.0.0
ipython==8.17.2